% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACHINE-class.R
\docType{class}
\name{MACHINE-class}
\alias{MACHINE-class}
\title{The MACHINE class}
\description{
The MACHINE object is used to store GWAS summary data, LD matrices, variables
useful for MCMC sampling, MACHINE hyper-parameters, posterior samples, and
credible sets. Objects can be created by CreateMACHINEObject.
}
\section{Slots}{

\describe{
\item{\code{M}}{Number of SNPs.}

\item{\code{P}}{Number of ancestries.}

\item{\code{N}}{A P-vector of GWAS sample sizes for all ancestries. For binary traits, 
'N' is the sum of the number of cases and the number of controls.}

\item{\code{Z}}{Z-score matrix with M rows and P columns. Missing values are set as NA.}

\item{\code{R}}{A list of P correlation matrices in \code{\link[Matrix]{dsCMatrix-class}} 
format. The size of each matrix matches the number of variants present in the
corresponding ancestry. The diagonal elements of each matrix are set as 0.}

\item{\code{SNP_ID}}{IDs of SNPs. A character vector of length M.}

\item{\code{POP_ID}}{IDs of ancestries. A character vector of length P.}

\item{\code{values}}{A list of P vectors. Each vector comprises eigenvalues of 
corresponding R (with diagonal elements as 1).}

\item{\code{tvectors}}{A list of P matrices. For each matrix, each column of its 
transpose is an eigenvector of corresponding R.}

\item{\code{lambda}}{A P-vector of hyper-parameters quantifying the discrepancy 
between z-scores and R for each ancestry.}

\item{\code{a}}{An M-vector of shape parameters for xi.}

\item{\code{b}}{Shape parameter for 1-h^2.}

\item{\code{c}}{An M-by-P matrix of shape parameters for eta.}

\item{\code{mcmc_samples}}{A list storing MCMC samples.}

\item{\code{mcmc_mean}}{A list storing posterior sample means.}

\item{\code{mcmc_sd}}{A list storing posterior sample standard errors.}

\item{\code{mcmc_quantile}}{A list storing posterior sample 2.5\%, 25\%, 50\%, 75\%, 
and 97.5\% quantiles.}

\item{\code{CL}}{An M-by-P-matrix of credible levels of variables for each ancestry.}

\item{\code{coverage}}{A number between 0 and 1 specifying the "coverage" of the 
credible sets.}

\item{\code{purity}}{A number between 0 and 1 specifying the minimum absolute 
correlation allowed in a credible set.}

\item{\code{CS}}{A list of identified credible sets in each ancestry.
See \code{\link{MACHINE_CS}}.}
}}

