% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACHINE-package.R
\docType{package}
\name{MACHINE}
\alias{MACHINE-package}
\alias{MACHINE}
\title{MACHINE package}
\description{
MACHINE: Multi-AnCestry Heritability INducEd Dirichlet decomposition
}
\details{
This package implements a multi-ancestry fine-mapping method based on the 
"Multi-AnCestry Heritability INducEd Dirichlet decomposition" (MACHINE) prior.
It is a novel multi-ancestry fine-mapping method that utilizes a continuous 
global-local shrinkage prior.
This method can be applied to both quantitative and binary traits.
An MCMC algorithm is employed to obtain samples from the posterior 
distribution.
This method can also provide "credible sets" of candidate causal variants, 
which are generally provided by fine-mapping methods based on 
discrete-mixture priors.
}
\references{
https://doi.org/10.1016/j.ajhg.2023.12.007
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xiangli428/MACHINE}
}

}
\author{
\strong{Maintainer}: Xiang Li \email{freddyl@connect.hku.hk}

Authors:
\itemize{
  \item Yan Dora Zhang \email{doraz@hku.hk}
}

}
\keyword{internal}
