% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACHINE_CS.R
\name{MACHINE_CS}
\alias{MACHINE_CS}
\title{Identify credible sets for MACHINE object.}
\usage{
CS = MACHINE_CS(MACHINE, coverage = 0.95, purity = 0.5)
}
\arguments{
\item{MACHINE}{A MACHINE object with MCMC samples.}

\item{coverage}{A number between 0 and 1 specifying the required coverage
of the credible sets.}

\item{purity}{A number between 0 and 1 specifying the minimum 
absolute correlation allowed in a credible set.}
}
\value{
A list of lists. Each list is for an ancestry and comprises the 
following elements:
\describe{
  \item{sets}{A list in which each element is a vector containing the
  indices of the variables in the CS.}
  \item{CL}{The credible level of each CS.}
  \item{purity}{The purity of each CS.}
}
}
\description{
A greedy search algorithm to identify all credible sets achieving target
"coverage" with minimum absolute correlation not less than "purity".
}
