% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMACHINEObject.R
\name{CreateMACHINEObject}
\alias{CreateMACHINEObject}
\title{Create a MACHINE object.}
\usage{
MACHINE = CreateMACHINEObject(Z,
                              R,
                              N,
                              SNP_ID = NULL,
                              POP_ID = NULL,
                              in_sample_LD = FALSE,
                              lambda = NULL,
                              R_eig = NULL,
                              a = 0.005,
                              b = NULL,
                              c = 0.2,
                              coverage = 0.95,
                              purity = 0.5,
                              tol = 1e-8)
}
\arguments{
\item{Z}{Z-score matrix with each row for a variant and each column for an
ancestry. Missing values are set as NA.}

\item{R}{A list of P correlation matrices that can be coerced to 
\code{\link[Matrix]{dsCMatrix-class}} format. The size of each matrix matches 
the number of variants present in the corresponding ancestry. The diagonal 
elements of each matrix will be coerced to zeros.}

\item{N}{GWAS sample size. For binary traits, 'N' is the sum of the
number of cases and the number of controls.}

\item{SNP_ID}{Identifiers of SNPs. The default is c("SNP_1", ...).}

\item{in_sample_LD}{Logical. Setting in_sample_LD = TRUE if in-sample LD
matrix is used.}

\item{lambda}{Hyper-parameter quantifying the discrepancy between z-scores
and R. If in_sample_LD == TRUE, lambda will be set as 0. Otherwise, if
lambda is not provided, lambda will be estimated.}

\item{R_eig}{An output of "eigen" function. A list with two components
"values" and "vector". If in_sample_LD == TRUE, don't need to provide it.
If provided, it will be used to create h2D2 object directly. Otherwise, 
it will be computed.}

\item{a}{Shape parameters for sigma^2. Either a positive real number or an
M-vector of positive real numbers.}

\item{b}{Shape parameter for 1-h^2. A positive real number. If not specified,
it will be estimated by an empirical Bayesian approach.}

\item{c}{Shape parameters for eta. Either a positive real number or an
M-by-P matrix of positive real numbers.}

\item{coverage}{A number between 0 and 1 specifying the required coverage
of the credible sets.}

\item{purity}{A number between 0 and 1 specifying the minimum 
absolute correlation allowed in a credible set.}

\item{tol}{Eigenvalues less than tol will be coerced to zeros.}
}
\value{
An h2D2 object. See \code{\link{h2D2-class}}.
The input z-scores will be modified to z / sqrt(1 + z^2/N - 1/N).
}
\description{
Create a MACHINE object from GWAS summary statistics and LD matrices of
multiple ancestries. GWAS z-scores, LD matrices, MACHINE hyper-parameters, 
MCMC samples, and outputs are stored.
}
